requirements.txt
torch>=1.8
tensorboard==2.3
numpy
scikit-learn
open3d>=0.8
Pillow
tqdm
MinkowskiEngine==0.5.4
